% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_tumopp}
\alias{cpp_tumopp}
\title{Run C++ simulation}
\usage{
cpp_tumopp(args)
}
\arguments{
\item{args}{string vector}
}
\value{
conf and population as strings
}
\description{
Run C++ simulation
}

