% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{evenness}
\alias{evenness}
\alias{shannon_index}
\alias{simpson_index}
\title{H / H_max}
\usage{
evenness(species)

shannon_index(species)

simpson_index(species)
}
\arguments{
\item{species}{a factor vector}
}
\value{
a number

a number

a number
}
\description{
H / H_max

Shannon index

Simpson index
}

