% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{extract_demography}
\alias{altered_params}
\alias{command_args}
\alias{exclusive_ancestors}
\alias{exclusive_ancestors_ss}
\alias{extract_ancestor}
\alias{extract_demography}
\title{extract demography from population data}
\usage{
extract_demography(grouped_df)

altered_params(conf)

exclusive_ancestors(population, size, origin = 1)

exclusive_ancestors_ss(snapshots, size)

extract_ancestor(genealogy, ids)

command_args()
}
\arguments{
\item{grouped_df}{a grouped_df}

\item{conf}{a data.frame}

\item{population}{a raw data.frame}

\item{size}{an integer}

\item{origin}{an integer}

\item{snapshots}{a data.frame}

\item{size}{an integer}

\item{string}{ancestors column}

\item{anc_ids}{exclusive ancestor ids}
}
\value{
a grouped data.frame

a string vector

an integer vector

an integer vector

an integer vector

a list
}
\description{
extract demography from population data

extract param names from conf

cell ids at the size

cell ids at the size (from snapshots)

split ancestor string and extract one matched

get commandline arguments
}
\examples{
extract_ancestor(c('1:2:9', '1:2:4:6'), c('2', '5', '6'))
}

