% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{first_ancestors}
\alias{first_ancestors}
\title{split ancestor string and extract an integer}
\usage{
first_ancestors(string, n)
}
\arguments{
\item{string}{ancestors column}

\item{n}{consider first n ancestors}
}
\value{
an integer vector
}
\description{
split ancestor string and extract an integer
}
\examples{
first_ancestors(c('1:2:9', '2:4:6'), 4)
}

