% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{trans_coord_hex_xy}
\alias{trans_coord_fcc}
\alias{trans_coord_hcc}
\alias{trans_coord_hex}
\alias{trans_coord_hex_xy}
\title{2D transformation into hexagonal grid}
\usage{
trans_coord_hex_xy(mtrx)

trans_coord_hcc(mtrx)

trans_coord_fcc(mtrx)

trans_coord_hex(mtrx)
}
\arguments{
\item{mtrx}{2D matrix}

\item{mtrx}{3D matrix}

\item{mtrx}{3D matrix}

\item{mtrx}{matrix}
}
\value{
transformed matrix

transformed matrix

transformed matrix

transformed matrix
}
\description{
2D transformation into hexagonal grid

3D transformation (hexagonal close packed)

3D transformation (face centered cubic, cubic close packed)

coordinate transformation
}

