% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{df2img}
\alias{detect_surface}
\alias{df2img}
\alias{filter_surface}
\alias{get_se}
\alias{img2df}
\title{Convert data.frame to cimg}
\usage{
df2img(mtrx)

img2df(img)

get_se(coord = c("moore", "neumann", "hex"), dimensions = 3)

filter_surface(img, se)

detect_surface(mtrx, se)
}
\arguments{
\item{mtrx}{a data.frame with (x, y, z) columns}

\item{img}{a cimg object}

\item{coord}{a string}

\item{dimensions}{an integer}

\item{se}{a cimg object}
}
\value{
cimg

a data.frame with (x, y, z) columns

cimg structuring element

cimg

tibble with surface column
}
\description{
Convert data.frame to cimg

Convert cimg to data.frame

Structuring element

Filter img by surface

extract cells on suface
}

