% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.R
\name{sprinkle_mutations}
\alias{sprinkle_mutations}
\alias{sample_stats}
\title{Make ms-like neutral variation matrix}
\usage{
sprinkle_mutations(population, nsam, mu = NULL, segsites = NULL)

sample_stats(msout, lowerbound = 0)
}
\arguments{
\item{population}{tibble of raw population}

\item{nsam}{number of cells to sample}

\item{mu}{mutation rate per cell division (ignored if segsites is given)}

\item{segsites}{number of segregating sites}

\item{msout}{tibble from `sprinkle_mutations()`}

\item{lowerbound}{to discard peripheral branches}
}
\value{
int matrix [nsam x segsites]

tibble
}
\description{
Make ms-like neutral variation matrix

Summary statistics of neutral mutations
}
