% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{save_serial_section}
\alias{save_serial_section}
\title{Write animation GIF for serial sections of 3D tumor}
\usage{
save_serial_section(.data, filename = "serial_section.gif", ...,
  width = 720, height = 640)
}
\arguments{
\item{.data}{tbl with extant cells}

\item{filename}{string}

\item{...}{passed to gglattice2d}

\item{width}{integer}

\item{height}{integer}
}
\description{
Write animation GIF for serial sections of 3D tumor
}
