% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_capture_rate}
\alias{plot_capture_rate}
\title{Plot capture_rate ~ nsam of biopsy}
\usage{
plot_capture_rate(data)
}
\arguments{
\item{data}{tbl from summarize_capture_rate()}
}
\description{
Plot capture_rate ~ nsam of biopsy
}
