% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_genealogy}
\alias{plot_bar_age}
\alias{plot_genealogy}
\title{Plot genealogy}
\usage{
plot_genealogy(.data, xmax = max(.data$ageend), colour = "dodgerblue")

plot_bar_age(.data, xmax = max(.data$ageend), alpha = 1, ...)
}
\arguments{
\item{.data}{tbl from layout_genealogy()}

\item{xmax}{numeric}

\item{colour}{character}

\item{alpha}{opacity [0, 1]}

\item{...}{passed to aes_()}
}
\value{
gg

gg
}
\description{
Plot genealogy

Plot age histogram
}

