% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-rgl.R
\name{plot_tumor3d}
\alias{plot_tumor3d}
\alias{snapshot_surface}
\title{plot tumor in 3d with rgl}
\usage{
plot_tumor3d(.tbl, colour = "clade", .palette = "Spectral",
  .reverse = FALSE, .min = NULL)

snapshot_surface(.tbl, filename = tempfile("rgl_", fileext = ".png"), ...)
}
\arguments{
\item{.tbl}{data.frame}

\item{colour}{column name to colorcode}

\item{.palette}{name for RColorBrewer::brewer.pal()}

\item{.reverse}{logical for order of color vector}

\item{.min}{minimum limit of axes}

\item{filename}{string}

\item{...}{passed to plot_tumor3d()}
}
\value{
filename
}
\description{
plot tumor in 3d with rgl

Shortcut of plot_tumor3d() %>% snapshot3d()
}
