% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{save_serial_section}
\alias{save_serial_section}
\title{Plot serial sections of 3D tumor}
\usage{
save_serial_section(.tbl, filename = "png/section_\%03d.png", scale = 6,
  dpi = 72, ...)
}
\arguments{
\item{.tbl}{tbl with extant cells}

\item{filename}{File name to create on disk.}

\item{scale}{Multiplicative scaling factor.}

\item{dpi}{Plot resolution. Applies only to raster output types.}

\item{...}{passed to plot_lattice2d}
}
\description{
Plot serial sections of 3D tumor
}
