% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius.R
\name{volume_function}
\alias{volume_function}
\alias{radius_function}
\title{Return volume function of specified coord and dims}
\usage{
volume_function(coord = "hex", dimensions = 3L)

radius_function(coord = "hex", dimensions = 3L)
}
\arguments{
\item{coord}{hex or not}

\item{dimensions}{integer}
}
\value{
function with radius argument

function with volume argument
}
\description{
Return volume function of specified coord and dims

Return radius function of specified coord and dims
}
