% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-combn.R
\name{combn_sample_ids}
\alias{combn_sample_ids}
\alias{summarize_capture_rate}
\alias{within_between_samples}
\title{Sample cells from a population}
\usage{
combn_sample_ids(ids, nsam = seq_along(ids))

summarize_capture_rate(combinations, population, threshold = 0.01)

within_between_samples(graph, regions)
}
\arguments{
\item{ids}{list of ID vectors}

\item{nsam}{number of regions to sample}

\item{combinations}{nested tibble from \code{combn_sample_ids()}}

\item{population}{tibble with descendants and id column}

\item{threshold}{lowerbound of detectable allele frequency}

\item{graph}{igraph}

\item{regions}{output of \code{sample_uniform_regions()}}
}
\value{
tibble
}
\description{
\code{combn_sample_ids} makes combinations of cell IDs for various number of samples

\code{summarize_capture_rate} calculates expected allele capture rate on various combinations of samples

\code{within_between_samples} summarizes branch lengths
}
