% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{sphere_radius}
\alias{sphere_radius}
\alias{sphere_volume}
\title{calculate radius from volume}
\usage{
sphere_radius(volume)

sphere_volume(radius)
}
\arguments{
\item{volume}{a numeric vector}

\item{radius}{a numeric vector}
}
\value{
radius a numeric vector

volume a numeric vector
}
\description{
calculate radius from volume

calculate volume from radius
}
\examples{
sphere_radius(42)
sphere_volume(42)
}

