% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{genetic_stats}
\alias{genetic_stats}
\alias{morphological_stats}
\title{Summary statistics of genealogy}
\usage{
genetic_stats(extant)

morphological_stats(extant)
}
\arguments{
\item{extant}{tibble}
}
\value{
tibble

tibble
}
\description{
Summary statistics of genealogy

Summary statistics of morphology
}
