% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumopp.R
\name{tumopp}
\alias{tumopp}
\title{Run C++ simulation}
\usage{
tumopp(args = character(0))
}
\arguments{
\item{args}{command line arguments as a string vector}
}
\value{
tibble
}
\description{
Run C++ simulation
}

