% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{maxabs}
\alias{maxabs}
\alias{trans_coord_fcc}
\alias{trans_coord_hcc}
\alias{trans_coord_hex}
\alias{trans_coord_hex_xy}
\title{get max value for plot limits}
\usage{
maxabs(mtrx)

trans_coord_hex_xy(mtrx)

trans_coord_hcc(mtrx)

trans_coord_fcc(mtrx)

trans_coord_hex(mtrx)
}
\arguments{
\item{mtrx}{a data.frame with (x, y, z) columns}
}
\value{
max(abs(x, y, z))

transformed matrix

transformed matrix

transformed matrix

transformed matrix
}
\description{
get max value for plot limits

2D transformation into hexagonal grid

3D transformation (hexagonal close packed)

3D transformation (face centered cubic, cubic close packed)

coordinate transformation
}
\examples{
maxabs(data.frame(x=2, y=-3, z=4))
}

