% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{max_abs_xyz}
\alias{dist_euclidean}
\alias{max_abs_xyz}
\alias{radius_function}
\alias{rotate}
\alias{trans_coord_fcc}
\alias{trans_coord_hcc}
\alias{trans_coord_hex}
\alias{trans_coord_hex_xy}
\alias{volume_function}
\title{get max value for plot limits}
\usage{
max_abs_xyz(.data)

trans_coord_hex_xy(.data)

trans_coord_hcc(.data)

trans_coord_fcc(.data)

trans_coord_hex(.data)

rotate(.data, theta, axis = c("z", "x", "y"))

dist_euclidean(.data, point = c(x = 0, y = 0, z = 0))

volume_function(coord = "hex", dimensions = 3L)

radius_function(coord = "hex", dimensions = 3L)
}
\arguments{
\item{.data}{data.frame with (x, y, z) columns}

\item{theta}{radian angle}

\item{axis}{a string}

\item{point}{named vector or tibble}

\item{coord}{hex or not}

\item{dimensions}{integer}
}
\value{
max(abs(x, y, z))

transformed matrix

transformed matrix

transformed matrix

transformed matrix

modified data.frame

numeric vector

function with radius argument

function with volume argument
}
\description{
get max value for plot limits

2D transformation into hexagonal grid

3D transformation (hexagonal close packed)

3D transformation (face centered cubic, cubic close packed)

2D/3D transformation into hexagonal grid

Rotate

Calculate distance from a specified cell

Return volume function of specified coord and dims

Return radius function of specified coord and dims
}
\examples{
max_abs_xyz(data.frame(x=2, y=-3, z=4))
}

