% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{max_abs_xyz}
\alias{max_abs_xyz}
\alias{rotate}
\alias{trans_coord_fcc}
\alias{trans_coord_hcc}
\alias{trans_coord_hex}
\alias{trans_coord_hex_xy}
\title{get max value for plot limits}
\usage{
max_abs_xyz(.data)

trans_coord_hex_xy(.data)

trans_coord_hcc(.data)

trans_coord_fcc(.data)

trans_coord_hex(.data)

rotate(.data, theta, axis = c("z", "x", "y"))
}
\arguments{
\item{.data}{data.frame with (x, y, z) columns}

\item{theta}{radian angle}

\item{axis}{a string}
}
\value{
max(abs(x, y, z))

transformed matrix

transformed matrix

transformed matrix

transformed matrix

modified data.frame
}
\description{
get max value for plot limits

2D transformation into hexagonal grid

3D transformation (hexagonal close packed)

3D transformation (face centered cubic, cubic close packed)

2D/3D transformation into hexagonal grid

Rotate
}
\examples{
max_abs_xyz(data.frame(x=2, y=-3, z=4))
}

