% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{altered_params}
\alias{altered_params}
\alias{count_descendants}
\alias{extract_demography}
\alias{filter_extant}
\alias{modify_population}
\alias{set_clades}
\title{extract param names from conf}
\usage{
altered_params(conf)

extract_demography(grouped_df)

filter_extant(raw_population)

modify_population(result, num_clades = 4L)

set_clades(raw_population, num_clades)

count_descendants(raw_population)
}
\arguments{
\item{conf}{a data.frame}

\item{grouped_df}{a grouped_df}

\item{raw_population}{tibble including ancestors}

\item{result}{a row of nested tibble}

\item{num_clades}{integer}
}
\value{
a string vector

a grouped data.frame

tibble

tibble

tibble

tibble with $id and $discendants
}
\description{
extract param names from conf

extract demography from population data

Filter extant cells

Modify population table

Add a column of ancestor ids by which branches are classified

Add a column of living descendants number
}

