% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{altered_params}
\alias{altered_params}
\alias{count_descendants}
\alias{extract_demography}
\alias{filter_extant}
\alias{modify_population}
\alias{set_clades}
\alias{set_id}
\title{extract param names from conf}
\usage{
altered_params(conf)

extract_demography(population)

filter_extant(population)

set_id(raw_population)

modify_population(result, num_clades = 4L)

set_clades(population, num_clades)

count_descendants(population)
}
\arguments{
\item{conf}{a data.frame}

\item{population}{tibble including ancestors}

\item{raw_population}{original tibble}

\item{result}{a row of nested tibble}

\item{num_clades}{integer}
}
\value{
a string vector

tibble

tibble

tibble

tibble

tibble

tibble with $id and $discendants
}
\description{
extract param names from conf

Extract demography from raw population data

Filter extant cells

Extract age and id from genealogy column

Modify population table

Add a column of ancestor ids by which branches are classified

Add a column of living descendants number
}

