% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{extract_demography}
\alias{altered_params}
\alias{extract_demography}
\alias{first_ancestors}
\title{extract demography from population data}
\usage{
extract_demography(grouped_df)

altered_params(conf)

first_ancestors(string, n)
}
\arguments{
\item{grouped_df}{a grouped_df}

\item{conf}{a data.frame}

\item{string}{ancestors column}

\item{n}{consider first n ancestors}
}
\value{
a grouped data.frame

a string vector

an integer vector
}
\description{
extract demography from population data

extract param names from conf

split ancestor string and extract an integer
}
\examples{
first_ancestors(c('1:2:9', '2:4:6'), 4)
}

