% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{extract_demography}
\alias{altered_params}
\alias{colorcode_survivors}
\alias{command_args}
\alias{exclusive_ancestors}
\alias{exclusive_ancestors_ss}
\alias{extract_ancestor}
\alias{extract_demography}
\title{extract demography from population data}
\usage{
extract_demography(grouped_df)

altered_params(conf)

exclusive_ancestors(population, size, origin = 1)

exclusive_ancestors_ss(snapshots, size)

extract_ancestor(genealogy, ids)

colorcode_survivors(raw_population, n = 4)

command_args()
}
\arguments{
\item{grouped_df}{a grouped_df}

\item{conf}{a data.frame}

\item{population}{a raw data.frame}

\item{size}{an integer}

\item{origin}{an integer}

\item{snapshots}{a data.frame}

\item{genealogy}{a string vector of colon-separated ancestors}

\item{ids}{a string vector of exclusive ancestor ids}

\item{raw_population}{a data.frame including ancestors}

\item{n}{an integer number of exclusive ancestors}

\item{size}{an integer}
}
\value{
a grouped data.frame

a string vector

an integer vector

an integer vector

a string vector of extracted ancestors

a data.frame

a list
}
\description{
extract demography from population data

extract param names from conf

cell ids that existed at the specified tumor size

cell ids that existed at the specified tumor size (from snapshots)

split ancestor string and extract one matched

E ancestor column

get commandline arguments
}
\examples{
extract_ancestor(c('1:2:9', '1:2:4:6'), c('2', '5', '6'))
}

