% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{altered_params}
\alias{altered_params}
\alias{colorcode_survivors}
\alias{exclusive_ancestors}
\alias{exclusive_ancestors_ss}
\alias{extract_ancestor}
\alias{extract_demography}
\title{extract param names from conf}
\usage{
altered_params(conf)

extract_demography(grouped_df)

exclusive_ancestors(raw_population, size, origin = 1)

exclusive_ancestors_ss(snapshots, size)

extract_ancestor(genealogy, ids)

colorcode_survivors(raw_population, n = 4)
}
\arguments{
\item{conf}{a data.frame}

\item{grouped_df}{a grouped_df}

\item{raw_population}{a data.frame including ancestors}

\item{size}{an integer}

\item{origin}{an integer}

\item{snapshots}{a data.frame}

\item{genealogy}{a string vector of colon-separated ancestors}

\item{ids}{a string vector of exclusive ancestor ids}

\item{n}{an integer number of exclusive ancestors}
}
\value{
a string vector

a grouped data.frame

an integer vector

an integer vector

a factor vector of extracted ancestors

a data.frame
}
\description{
extract param names from conf

extract demography from population data

cell ids that existed at the specified tumor size

cell ids that existed at the specified tumor size (from snapshots)

split ancestor string and extract one matched

E ancestor column
}
\examples{
extract_ancestor(c('1:2:9', '1:2:4:6'), c('2', '5', '6'))
}

