% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{extract_demography}
\alias{altered_params}
\alias{command_args}
\alias{exclusive_ancestors}
\alias{extract_demography}
\alias{filter_ancestors}
\alias{first_ancestors}
\title{extract demography from population data}
\usage{
extract_demography(grouped_df)

altered_params(conf)

exclusive_ancestors(snapshots, size)

first_ancestors(string, anc_ids)

filter_ancestors(mtrx, anc_ids)

command_args()
}
\arguments{
\item{grouped_df}{a grouped_df}

\item{conf}{a data.frame}

\item{snapshots}{a data.frame}

\item{size}{an integer}

\item{string}{ancestors column}

\item{anc_ids}{an integer vector}

\item{mtrx}{a data.frame}

\item{anc_ids}{an integer vector}
}
\value{
a grouped data.frame

a string vector

an integer vector

an integer vector

a modified data.frame

a list
}
\description{
extract demography from population data

extract param names from conf

cell ids at the size

split ancestor string and extract an integer

filter first ancesters for coloring

get commandline arguments
}
\examples{
first_ancestors(c('1:2:9', '1:2:4:6'), c(2, 5, 6))
}

