% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{altered_params}
\alias{altered_params}
\alias{count_descendants}
\alias{exclusive_ancestors}
\alias{exclusive_ancestors_ss}
\alias{extract_ancestor}
\alias{extract_demography}
\alias{extract_survivors}
\title{extract param names from conf}
\usage{
altered_params(conf)

extract_demography(grouped_df)

exclusive_ancestors(raw_population, size, origin = 1)

exclusive_ancestors_ss(snapshots, size)

extract_ancestor(genealogy, ids)

extract_survivors(result, n = 4)

count_descendants(raw_population)
}
\arguments{
\item{conf}{a data.frame}

\item{grouped_df}{a grouped_df}

\item{raw_population}{a data.frame including ancestors}

\item{size}{an integer}

\item{origin}{an integer}

\item{snapshots}{a data.frame}

\item{genealogy}{list of integer vectors}

\item{ids}{integers}

\item{result}{a row of a nested tibble of conf and list(pop)}

\item{n}{an integer number of exclusive ancestors}
}
\value{
a string vector

a grouped data.frame

an integer vector

an integer vector

factor vector of extracted ancestors

a data.frame

tibble with $id and $discendants
}
\description{
extract param names from conf

extract demography from population data

cell ids that existed at the specified tumor size

cell ids that existed at the specified tumor size (from snapshots)

split ancestor string and extract one matched

Extract survivors and add some columns

Add descendants column
}

