% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{math_score}
\alias{math_score}
\alias{count_descendants}
\alias{descfreqs}
\alias{detectable_ancestors}
\title{Calculate MATH (mutant-allele tumor heterogeneity) score from variant allele frequencies}
\usage{
math_score(x, constant = 1.4826, na.rm = FALSE)

count_descendants(graph, nodes)

descfreqs(descendants)

detectable_ancestors(population, threshold = 0.01)
}
\arguments{
\item{x}{a numeric vector.}

\item{constant}{scale factor.}

\item{na.rm}{if \code{TRUE} then \code{NA} values are stripped
    from \code{x} before computation takes place.}

\item{graph}{The input graph.}

\item{nodes}{The vertices for which the calculation is performed.}

\item{descendants}{column of raw population including ancestors}

\item{population}{tibble with descendants and id column}

\item{threshold}{lowerbound of detectable allele frequency}
}
\value{
numeric

integer

numeric vector applicable to math_score

integer IDs
}
\description{
Calculate MATH (mutant-allele tumor heterogeneity) score from variant allele frequencies

Count number of extant descendants

Translate descendant numbers to frequencies

Extract ancestors whose descendants frequency is above threshold
}
