% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{math_score}
\alias{descendants_fs}
\alias{math_score}
\title{Calculate MATH (mutant-allele tumor heterogeneity) score from variant allele freqneucies}
\usage{
math_score(x, constant = 1.4826, na.rm = FALSE)

descendants_fs(descendants, lowerbound = 0)
}
\arguments{
\item{x}{a numeric vector.}

\item{constant}{scale factor.}

\item{na.rm}{if \code{TRUE} then \code{NA} values are stripped
    from \code{x} before computation takes place.}

\item{descendants}{column of raw population including ancestors}

\item{lowerbound}{to discard peripheral branches}
}
\value{
numeric

numeric vector
}
\description{
Calculate MATH (mutant-allele tumor heterogeneity) score from variant allele freqneucies

Make true spectrum from which allele frequencies are generated
}

