% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{math_score}
\alias{math_score}
\title{Calculate MATH score from variant allele freqneucies}
\usage{
math_score(freqs)
}
\arguments{
\item{freqs}{numeric vector}
}
\value{
numeric
}
\description{
Calculate MATH score from variant allele freqneucies
}

