% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{math_score}
\alias{math_score}
\alias{descfreqs}
\title{Calculate MATH (mutant-allele tumor heterogeneity) score from variant allele frequencies}
\usage{
math_score(x, constant = 1.4826, na.rm = FALSE)

descfreqs(descendants, lowerbound = 0)
}
\arguments{
\item{x}{a numeric vector.}

\item{constant}{scale factor.}

\item{na.rm}{if \code{TRUE} then \code{NA} values are stripped
    from \code{x} before computation takes place.}

\item{descendants}{column of raw population including ancestors}

\item{lowerbound}{to discard peripheral branches}
}
\value{
numeric

numeric vector applicable to math_score
}
\description{
Calculate MATH (mutant-allele tumor heterogeneity) score from variant allele frequencies

Translate descendant numbers to frequencies
}
