% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{make_edgelist}
\alias{ggplot_genealogy}
\alias{layout_genealogy}
\alias{make_edgelist}
\title{Make parent-child list from raw population tibble}
\usage{
make_edgelist(population)

layout_genealogy(edgelist)

ggplot_genealogy(.data)
}
\arguments{
\item{population}{tibble}

\item{edgelist}{tibble from make_edgelist()}

\item{.data}{tibble from layout_genealogy()}
}
\value{
tibble

tibble

gg
}
\description{
Make parent-child list from raw population tibble

Set coordinates of nodes and edges for plotting

Plot genealogy
}

