% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{make_igraph}
\alias{make_igraph}
\alias{mean_branch_length}
\alias{layout_genealogy}
\title{Make igraph from raw population tbl}
\usage{
make_igraph(population)

mean_branch_length(graph, from = igraph::V(graph), to = from)

layout_genealogy(population)
}
\arguments{
\item{population}{tbl}

\item{graph}{igraph}

\item{from, to}{igraph vertices}
}
\value{
tibble

numeric

tibble
}
\description{
Make igraph from raw population tbl

Mean branch length within/between sub-graphs

Set coordinates of nodes and edges for plotting
}
