% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{make_edgelist}
\alias{make_edgelist}
\title{cell ids that existed at the specified tumor size}
\usage{
make_edgelist(population)
}
\arguments{
\item{population}{a raw data.frame}
}
\value{
a data.frame
}
\description{
cell ids that existed at the specified tumor size
}

