% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{make_edgelist}
\alias{make_edgelist}
\alias{make_igraph}
\alias{mean_branch_length}
\alias{mean_branch_length.igraph}
\alias{layout_genealogy}
\title{Make parent-child list from raw population tbl or igraph}
\usage{
make_edgelist(obj)

make_igraph(population, ids = integer(0L))

mean_branch_length(population, from = integer(0L), to = from)

mean_branch_length.igraph(graph, from = igraph::V(graph), to = from)

layout_genealogy(population)
}
\arguments{
\item{obj}{population tbl or igraph}

\item{population}{tbl}

\item{ids}{integer vector}

\item{from, to}{node id}

\item{graph}{igraph}
}
\value{
tibble

tibble

numeric

numeric

tibble
}
\description{
Make parent-child list from raw population tbl or igraph

Make igraph from raw population tbl

Mean branch length within/between sub-graphs

Mean branch length within/between sub-graphs

Set coordinates of nodes and edges for plotting
}
