% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{make_edgelist}
\alias{layout_genealogy}
\alias{make_edgelist}
\alias{make_igraph}
\alias{mean_branch_length}
\alias{mean_branch_length.igraph}
\alias{plot_age_histogram}
\alias{plot_genealogy}
\title{Make parent-child list from raw population tbl or igraph}
\usage{
make_edgelist(obj)

make_igraph(population, ids = integer(0L))

mean_branch_length(population, from = integer(0L), to = from)

mean_branch_length.igraph(graph, from = igraph::V(graph), to = from)

layout_genealogy(graph)

plot_genealogy(.data, xmax = max(.data$ageend), colour = "dodgerblue")

plot_age_histogram(.data, xmax = max(.data$ageend), alpha = 1, ...)
}
\arguments{
\item{obj}{population tbl or igraph}

\item{population}{tbl}

\item{ids}{integer vector}

\item{from, to}{node id}

\item{graph}{igraph}

\item{.data}{tbl from layout_genealogy()}

\item{xmax}{numeric}

\item{colour}{character}

\item{alpha}{opacity}

\item{...}{passed to aes_()}
}
\value{
tibble

tibble

numeric

numeric

tibble

gg

gg
}
\description{
Make parent-child list from raw population tbl or igraph

Make igraph from raw population tbl

Mean branch length within/between sub-graphs

Mean branch length within/between sub-graphs

Set coordinates of nodes and edges for plotting

Plot genealogy

Plot age histogram
}

