% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{make_edgelist}
\alias{age_histogram}
\alias{fst_HBK}
\alias{fst_HSM}
\alias{ggplot_genealogy}
\alias{layout_genealogy}
\alias{make_edgelist}
\alias{mean_branch_length}
\title{Make parent-child list from raw population tibble}
\usage{
make_edgelist(population)

mean_branch_length(graph, from = igraph::V(graph), to = from)

fst_HSM(within, between)

fst_HBK(within, between, n = 2)

layout_genealogy(edgelist)

ggplot_genealogy(.data, label = "", xmax = 20, hist = TRUE)

age_histogram(.data, age_lim)
}
\arguments{
\item{population}{tibble}

\item{graph}{igraph}

\item{from, to}{node names}

\item{within, between}{mean branch length or diversity}

\item{n}{number of subpopulations}

\item{edgelist}{tibble from make_edgelist()}

\item{.data}{tibble from layout_genealogy()}

\item{label}{character or expression}

\item{xmax}{numeric}

\item{hist}{logical: whether add age_histogram() or not.}

\item{age_lim}{numeric vector}
}
\value{
tibble

numeric

numeric

numeric

tibble

gg

gg
}
\description{
Make parent-child list from raw population tibble

Mean branch length within/between sub-graphs

Calculate Fst by Hudson, Slatkin, and Maddison (1992)

Calculate Kst by Hudson, Boos, and Kaplan (1992)

Set coordinates of nodes and edges for plotting

Plot genealogy

Plot age histogram
}

