% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{maxabs}
\alias{maxabs}
\title{get max value for plot limits}
\usage{
maxabs(mtrx)
}
\arguments{
\item{mtrx}{data.frame with (x, y, z) columns}
}
\value{
max(abs(x, y, z))
}
\description{
get max value for plot limits
}
\examples{
maxabs(data.frame(x=2, y=-3, z=4))
}

