% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{df2img}
\alias{df2img}
\alias{img2df}
\alias{get_se}
\alias{filter_surface}
\alias{detect_surface}
\title{Convert data.frame to binary array}
\usage{
df2img(mtrx)

img2df(img)

get_se(coord = c("moore", "neumann", "hex"), dimensions = 3L)

filter_surface(img, se)

detect_surface(mtrx, se)
}
\arguments{
\item{mtrx}{a data.frame with (x, y, z) columns}

\item{img}{a binary array}

\item{coord}{a string}

\item{dimensions}{an integer}

\item{se}{structuring element as a binary array}
}
\value{
binary array

a data.frame with (x, y, z) columns

structuring element as a binary array

a binary array

tibble with surface column
}
\description{
Convert data.frame to binary array

Convert binary array to data.frame

Structuring element (kernel)

Filter img by surface

extract cells on suface
}
