% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_genealogy}
\alias{plot_genealogy}
\alias{plot_bar_age}
\title{Plot genealogy}
\usage{
plot_genealogy(.tbl, xmax = max(.tbl$ageend), colour = "dodgerblue")

plot_bar_age(.tbl, xmax = max(.tbl$ageend), alpha = 1, ...)
}
\arguments{
\item{.tbl}{tbl from layout_genealogy()}

\item{xmax}{numeric}

\item{colour}{character}

\item{alpha}{opacity \code{[0, 1]}}

\item{...}{passed to aes_()}
}
\value{
gg

gg
}
\description{
Plot genealogy

Plot age histogram
}
