% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gglattice2D}
\alias{gglattice2D}
\title{ggplot for 2D lattice}
\usage{
gglattice2D(.data, colour = "ancestors", limit = maxabs(.data))
}
\arguments{
\item{.data}{a data.frame}

\item{colour}{a column name for colour}

\item{limit}{for value range}
}
\description{
ggplot for 2D lattice
}

