% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggfreqspec}
\alias{ggfreqspec}
\alias{gglattice2d}
\alias{plot_tumor3d}
\alias{save_serial_section}
\title{ggplot for frequency spectrum}
\usage{
ggfreqspec(freqs)

gglattice2d(.data, colour = "clade", alpha = 0.66, size = 1,
  limit = max_abs_xyz(.data))

save_serial_section(.data, filename = "serial_section.gif", ...,
  width = 720, height = 640)

plot_tumor3d(.data, colour = "clade", .palette = "Spectral")
}
\arguments{
\item{freqs}{numeric vector}

\item{.data}{data.frame}

\item{colour}{column name to colorcode}

\item{alpha}{opacity [0, 1]}

\item{size}{relative size of points}

\item{limit}{for value range}

\item{filename}{string}

\item{...}{passed to gglattice2d}

\item{width}{integer}

\item{height}{integer}

\item{.palette}{name for RColorBrewer::brewer.pal()}
}
\description{
ggplot for frequency spectrum

ggplot for 2D lattice

Write animation GIF for serial sections of 3D tumor

plot tumor in 3d with rgl
}

