% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggfreqspec}
\alias{ggfreqspec}
\alias{gglattice2d}
\alias{plot_tumor3d}
\alias{save_serial_section}
\title{ggplot for frequency spectrum}
\usage{
ggfreqspec(freqs)

gglattice2d(.data, .color = "clade", .palette = "Spectral",
  limit = maxabs(.data))

save_serial_section(filename = "serial_section.gif", .data, width = 720,
  height = 640)

plot_tumor3d(.data, .color = "clade", .palette = "Spectral")
}
\arguments{
\item{freqs}{numeric vector}

\item{.data}{a data.frame}

\item{.color}{a column name to colorcode}

\item{.palette}{name for RColorBrewer::brewer.pal()}

\item{limit}{for value range}

\item{filename}{a string}

\item{width}{an integer}

\item{height}{an integer}
}
\description{
ggplot for frequency spectrum

ggplot for 2D lattice

Write animation GIF for serial sections of 3D tumor

plot tumor in 3d with rgl
}

