% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{modify_population}
\alias{modify_population}
\alias{filter_extant}
\alias{filter_connected}
\alias{set_id}
\alias{set_clades}
\alias{count_descendants}
\alias{extract_demography}
\title{Modify population table}
\usage{
modify_population(result, num_clades = 4L)

filter_extant(population)

filter_connected(population, ids)

set_id(population)

set_clades(population, num_clades)

count_descendants(population)

extract_demography(population)
}
\arguments{
\item{result}{a row of nested tibble}

\item{num_clades}{integer}

\item{population}{tibble including ancestors}

\item{ids}{integer vector of extant/sample cells}
}
\value{
tibble

tibble

tibble

tibble

tibble

tibble with $id and $discendants

tibble
}
\description{
Modify population table

Filter extant cells

Filter connected cells

Extract age and id from genealogy column

Add a column of ancestor ids by which branches are classified

Add a column of living descendants number

Extract demography from raw population data
}
