% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{.read_conf}
\alias{.read_conf}
\alias{read_conf}
\alias{read_population}
\alias{read_snapshots}
\title{read a config file}
\usage{
.read_conf(filename)

read_conf(indir = ".")

read_snapshots(conf)

read_population(conf, params = NULL)
}
\arguments{
\item{filename}{a string}

\item{indir}{a string vector}

\item{conf}{a data.frame}

\item{params}{a string vector; columns to be preserved}

\item{conf}{a data.frame}
}
\value{
a data.frame

a data.frame

a grouped data.frame

a grouped data.frame
}
\description{
read a config file

read config files

read snapshots

read population
}

