% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_conf}
\alias{modify_population}
\alias{read_conf}
\alias{read_population}
\alias{read_snapshots}
\title{read config files}
\usage{
read_conf(indirs = ".")

read_population(conf)

modify_population(raw_population)

read_snapshots(conf)
}
\arguments{
\item{indirs}{a string vector}

\item{conf}{tibble}

\item{raw_population}{tibble including ancestors}
}
\value{
tibble

tibble

tibble

a grouped data.frame
}
\description{
read config files

read a population

Modify population table

read snapshots
}

