% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_confs}
\alias{read_confs}
\alias{read_populations}
\alias{read_results}
\alias{read_snapshots}
\title{Read config files}
\usage{
read_confs(indirs = getwd())

read_populations(indirs = getwd())

read_results(indirs = getwd())

read_snapshots(indirs = getwd())
}
\arguments{
\item{indirs}{a string vector}
}
\value{
tibble

list of tibbles

nested tibble

a grouped data.frame
}
\description{
Read config files

Read populations

Read confs and populations as a nested tibble

read snapshots
}
