% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_conf}
\alias{read_conf}
\alias{read_population}
\alias{read_snapshots}
\title{read config files}
\usage{
read_conf(indir = ".")

read_population(conf, params = NULL)

read_snapshots(conf)
}
\arguments{
\item{indir}{a string vector}

\item{conf}{a data.frame}

\item{params}{a string vector; columns to be preserved}
}
\value{
a data.frame

a grouped data.frame

a grouped data.frame
}
\description{
read config files

read population

read snapshots
}

