% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_conf}
\alias{read_conf}
\alias{read_population}
\alias{read_results}
\alias{read_snapshots}
\title{read config files}
\usage{
read_conf(indirs = ".")

read_population(conf)

read_results(indirs = ".")

read_snapshots(conf)
}
\arguments{
\item{indirs}{a string vector}

\item{conf}{tibble}

\item{indirs}{a string vector}
}
\value{
tibble

tibble

nested tibble

a grouped data.frame
}
\description{
read config files

read a population

Read results

read snapshots
}

