% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_conf}
\alias{command_args}
\alias{read_conf}
\alias{read_population}
\alias{read_snapshots}
\title{read config files}
\usage{
read_conf(indirs = ".")

read_population(conf)

read_snapshots(conf)

command_args()
}
\arguments{
\item{indirs}{a string vector}

\item{conf}{a data.frame}
}
\value{
a data.frame

a data.frame

a grouped data.frame

a list
}
\description{
read config files

read a population

read snapshots

get commandline arguments
}

