% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_bulk}
\alias{sample_bulk}
\alias{make_ancestor_union}
\alias{sprinkle_mutations}
\title{Sample bulk of cells near the specified coord}
\usage{
sample_bulk(extant, center, size = 100L)

make_ancestor_union(population, samples, origin = 1L)

sprinkle_mutations(population, nsam, mu = NULL, segsites = NULL)
}
\arguments{
\item{extant}{tibble of extant cells}

\item{center}{named (x, y, z) vector, list, or tibble}

\item{size}{number of cells}

\item{population}{tibble of raw population}

\item{samples}{vector of IDs}

\item{origin}{ID of common ancestor}

\item{nsam}{number of cells to sample}

\item{mu}{mutation rate per cell division (ignored if segsites is given)}

\item{segsites}{number of segregating sites}
}
\value{
tibble

union of IDs

int matrix [nsam x segsites]
}
\description{
Sample bulk of cells near the specified coord

Traceback ancestors and calculate union of the IDs

Make ms-like neutral variation matrix
}
