% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_bulk}
\alias{sample_bulk}
\alias{sample_random_regions}
\alias{within_between_samples}
\alias{make_ancestor_union}
\alias{sprinkle_mutations}
\title{Sample bulk of cells near the specified coord}
\usage{
sample_bulk(extant, center, size = 100L)

sample_random_regions(extant, nsam = 2L, ncell = 10L)

within_between_samples(graph, regions)

make_ancestor_union(graph, nodes)

sprinkle_mutations(population, nsam, mu = NULL, segsites = NULL)
}
\arguments{
\item{extant}{tibble of extant cells}

\item{center}{named (x, y, z) vector, list, or tibble}

\item{size}{number of cells}

\item{nsam}{number of cells to sample}

\item{ncell}{number of cells per specimen}

\item{graph}{igraph}

\item{regions}{output of sample_random_regions()}

\item{nodes}{igraph vertices}

\item{population}{tibble of raw population}

\item{mu}{mutation rate per cell division (ignored if segsites is given)}

\item{segsites}{number of segregating sites}
}
\value{
tibble

nested tibble

tibble

union of IDs

int matrix [nsam x segsites]
}
\description{
Sample bulk of cells near the specified coord

Sample multiple regions from extant cells

Sample specimens and measure mean branch lengths

Traceback ancestors and calculate union of the IDs

Make ms-like neutral variation matrix
}
