% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_bulk}
\alias{sample_bulk}
\alias{sprinkle_mutations}
\title{Sample bulk of cells near the specified coord}
\usage{
sample_bulk(population, center, size = 100L)

sprinkle_mutations(genealogy, mu = NULL, segsites = NULL)
}
\arguments{
\item{population}{tibble of extant cells}

\item{center}{named (x, y, z) vector, list, or tibble}

\item{size}{number of cells}

\item{genealogy}{list of int vectors}

\item{mu}{mutation rate per cell division (ignored if segsites is given)}

\item{segsites}{number of segregating sites}
}
\value{
tibble

int matrix [nsam x segsites]
}
\description{
Sample bulk of cells near the specified coord

Make ms-like neutral variation matrix from genealogy column of extant cells
}
