% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_bulk}
\alias{sample_bulk}
\alias{make_samples}
\title{Sample bulk of cells near the specified coord}
\usage{
sample_bulk(population, center, size = 100L)

make_samples(genealogy, nsam = length(genealogy), mu = NULL,
  segsites = NULL)
}
\arguments{
\item{population}{tibble of extant cells}

\item{center}{named (x, y, z) vector, list, or tibble}

\item{size}{number of cells}

\item{genealogy}{list of int vectors}

\item{nsam}{number of samples}

\item{mu}{mutation rate per cell division (ignored if segsites is given)}

\item{segsites}{number of segregating sites}
}
\value{
tibble

int matrix [nsam x segsites]
}
\description{
Sample bulk of cells near the specified coord

Make ms-like matrix from genealogy column of population
}
