% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{make_samples}
\alias{make_samples}
\title{Make ms-like matrix from survivors tibble}
\usage{
make_samples(survivors, nsam = nrow(survivors), mu = 0.1)
}
\arguments{
\item{survivors}{tibble}

\item{nsam}{number of samples}

\item{mu}{mutation rate per cell division}
}
\value{
logical tibble
}
\description{
Make ms-like matrix from survivors tibble
}

