% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R, R/summarize.R
\name{sample_stats}
\alias{sample_stats}
\alias{genetic_stats}
\alias{morphological_stats}
\title{Summary statistics of neutral mutations}
\usage{
sample_stats(msout, lowerbound = 0)

genetic_stats(extant)

morphological_stats(extant)
}
\arguments{
\item{msout}{tibble from `sprinkle_mutations()`}

\item{lowerbound}{to discard peripheral branches}

\item{extant}{tibble}
}
\value{
tibble

tibble

tibble
}
\description{
Summary statistics of neutral mutations

Summary statistics of genealogy

Summary statistics of morphology
}
