% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R, R/summarize.R
\name{sample_stats}
\alias{genetic_stats}
\alias{morphological_stats}
\alias{sample_stats}
\alias{within_between_samples}
\title{Summary statistics of neutral mutations}
\usage{
sample_stats(msout, lowerbound = 0)

genetic_stats(extant)

within_between_samples(population, ncell = 100L, npair = 1L)

morphological_stats(extant)
}
\arguments{
\item{msout}{tibble from `make_samples()`}

\item{lowerbound}{to discard peripheral branches}

\item{extant}{tibble}

\item{population}{tibble}

\item{ncell}{number of cells per specimen}

\item{npair}{number of pairs and output rows}
}
\value{
tibble

tibble

tibble

tibble
}
\description{
Summary statistics of neutral mutations

Summary statistics of genealogy

Sample two specimens and measure mean branch lengths

Summary statistics of morphology
}

