% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/tumopp.R
\docType{package}
\name{cpp_tumopp}
\alias{cpp_tumopp}
\alias{cpp_tumopp_ms}
\alias{tumopp}
\alias{tumopp-package}
\alias{tumopp}
\alias{ms}
\alias{make_args}
\title{Run C++ simulation}
\usage{
cpp_tumopp(args, npair = 0L)

cpp_tumopp_ms(nsam, args)

tumopp(args = character(0L), npair = 0L, nsam = 0L)

ms(nsam = 20L, args = character(0L))

make_args(alt, const = NULL, nreps = 1L)
}
\arguments{
\item{args}{command line arguments as a string vector or list of strings}

\item{npair}{number of samples to measure genetic and physical distance}

\item{nsam}{number of samples for ms-like output}

\item{alt}{named list of altered arguments}

\item{const}{unnamed vector of constant arguments}

\item{nreps}{number of repeats}
}
\value{
conf and population as strings

nested tibble

list of character vectors
}
\description{
Run C++ simulation

R interface of tumopp, tumor growth simulator in C++

Run C++ simulation

Make argument list for tumopp()
}
