% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumopp.R
\name{tumopp}
\alias{make_args}
\alias{tumopp}
\title{Run C++ simulation}
\usage{
tumopp(args = character(0))

make_args(alt, const = NULL, nreps = 1)
}
\arguments{
\item{args}{command line arguments as a string vector}

\item{alt}{named list of altered arguments}

\item{const}{unnamed vector of constant arguments}

\item{nreps}{number of repeats}
}
\value{
tibble

list of character vectors
}
\description{
Run C++ simulation

Make argment list for tumopp()
}

