% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/tumopp.R
\name{cpp_tumopp}
\alias{cpp_tumopp}
\alias{make_args}
\alias{tumopp}
\title{Run C++ simulation}
\usage{
cpp_tumopp(args, nsam = 0L)

tumopp(args = character(0L), nsam = 0L)

make_args(alt, const = NULL, nreps = 1L)
}
\arguments{
\item{args}{command line arguments as a string vector or list of strings}

\item{nsam}{number of samples to measure genetic and physical distance}

\item{alt}{named list of altered arguments}

\item{const}{unnamed vector of constant arguments}

\item{nreps}{number of repeats}
}
\value{
conf and population as strings

nested tibble

list of character vectors
}
\description{
Run C++ simulation

Run C++ simulation

Make argment list for tumopp()
}

